Split commands in separate modules
By importing the functions in ``__init__.py``,
the bot can import the commands as before without changing its behaviour
