+++ src/main.py
@@ -1,12 +1,14 @@
+import datetime
 import importlib
 import logging
 import sys
 
 import dotenv
+import pytz
 
 import db
 import settings
-import shuffle_commands
+import utils
 from koduck import Koduck, KoduckContext
 
 
@@ -67,8 +69,47 @@ async def refresh_commands(context: KoduckContext) -> None:
             )
         return
 
-    if errors:
-        print(errors)
+    for e in errors:
+        print(e)
+
+
+current_day = -1
+current_week = -1
+
+
+async def background_task(koduck: Koduck) -> None:
+    global current_day
+    global current_week
+    current_time = datetime.datetime.now(tz=pytz.timezone("Etc/GMT+6"))
+
+    if current_day == -1 or current_week == -1:
+        current_day = current_time.day
+        current_week = utils.get_current_week()
+        return
+
+    if current_time.day == current_day:
+        return
+
+    current_day = current_time.day
+    current_week2 = utils.get_current_week()
+    week_changed, current_week = current_week2 != current_week, current_week2
+    event_pokemon = utils.get_current_event_pokemon()
+    for reminder in db.get_reminders():
+        reminder_strings: list[str] = []
+        if week_changed and current_week in reminder.weeks:
+            reminder_strings.append(settings.message_reminder_week.format(current_week))
+        reminder_strings.extend(
+            settings.message_reminder_pokemon.format(ep)
+            for ep in event_pokemon
+            if ep in reminder.pokemon
+        )
+        if reminder_strings:
+            the_user = await koduck.client.fetch_user(reminder.user_id)
+            await the_user.send(
+                content=settings.message_reminder_header.format(
+                    reminder.user_id, "\n".join(reminder_strings)
+                )
+            )
 
 
 def main() -> None:
@@ -78,7 +119,7 @@ def main() -> None:
         print("Token not configured")
         sys.exit(1)
 
-    settings.background_task = shuffle_commands.background_task
+    settings.background_task = background_task
 
     koduck = Koduck()
     koduck.add_command("refreshcommands", refresh_commands, "prefix", 3)
